Infrastruktura/functions/fa-submit/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 
azure-identity==1.17.1
azure-servicebus==7.14.2
azure-functions
msal
psycopg2-binary==2.9.9
PyJWT==2.9.0
